added admin page & some UI fix
