+++ .env.example
@@ -15,7 +15,6 @@ BOT_TOKEN = "123"
 GUILD_ID = "123"
 LOCAL_CMS_PROVIDER = true
 CACHE_EXPIRE_S = 10
-
-ADMINS = "Random,example@gmail.com"
+NEXT_PUBLIC_ADMINS = "Random,example@gmail.com,test@gmail.com"
 NEXT_PUBLIC_DISABLE_FEATURES = "featurea,featureb,featurec"
 REDIS_URL=
