+++ src/app/admin/content/[...courseId]/page.tsx
@@ -44,7 +44,7 @@ export default async function UpdateCourseContent({
   return (
     <div className="mx-auto max-w-screen-xl justify-between p-4 text-black dark:text-white">
       {course?.title}
-      <div className="font-bold md:text-5xl lg:text-6xl">Content</div>
+      <div className="my-2 font-bold md:text-5xl lg:text-6xl">Content</div>
       <AddContent
         courseId={parseInt(courseId, 10)}
         parentContentId={parseFloat(rest[rest.length - 1])}
