+++ src/components/profile-menu/ProfileDropdown.tsx
@@ -19,8 +19,11 @@ import {
 import ExternalLinks from './ExternalLinks';
 import { signOut } from 'next-auth/react';
 import { SelectTheme } from './SelectTheme';
+import { useRouter } from 'next/navigation';
+import { FaChessKing } from 'react-icons/fa';
+import { Session } from 'next-auth';
 
-const ProfileDropdown = () => {
+const ProfileDropdown = ({ Isession }: { Isession: Session }) => {
   const menuItemLinks = [
     {
       href: '/history',
@@ -38,6 +41,7 @@ const ProfileDropdown = () => {
       label: 'Questions',
     },
   ];
+  const router = useRouter();
 
   return (
     <DropdownMenu modal={false}>
@@ -47,6 +51,19 @@ const ProfileDropdown = () => {
         </div>
       </DropdownMenuTrigger>
       <DropdownMenuContent className="mr-3 mt-2 w-56 shadow-2xl">
+        {Isession &&
+          process.env.NEXT_PUBLIC_ADMINS?.split(',').includes(
+            Isession?.user?.email ?? '',
+          ) && (
+            <DropdownMenuItem
+              onClick={() => {
+                router.push('/admin');
+              }}
+            >
+              <FaChessKing className="mr-2 h-4 w-4" />
+              <span>Admin</span>
+            </DropdownMenuItem>
+          )}
         <DropdownMenuGroup>
           {menuItemLinks.map(({ href, label, icon }) => (
             <Link href={href} key={href}>
